1.将操作类型修改到配置文件 2.方圆智控平台接口新加状态返回值 3.修改APP获取土地列表加入地块 3.删除土地加入如果是地块的逻辑判断

+++ ruoyi-service/fangyuan-api/src/main/java/com/ruoyi/fangyuanapi/service/impl/DbLandServiceImpl.java
@@ -96,6 +96,7 @@ public class DbLandServiceImpl implements IDbLandService
     @Override
     public int deleteDbLandByIds(String ids)
     {
+
         return dbLandMapper.deleteDbLandByIds(Convert.toStrArray(ids));
     }
 
@@ -105,8 +106,13 @@ public class DbLandServiceImpl implements IDbLandService
      * @param landId 土地ID
      * @return 结果
      */
+    @Transactional
     public int deleteDbLandById(Long landId)
     {
+        Integer i = dbLandMapper.selectDbLandBySiteId(landId);
+        if (i >0){
+            return 0;
+        }
         return dbLandMapper.deleteDbLandById(landId);
     }
